src/main/java/ladder/domain/Line.java
package ladder.domain;


import java.util.List;
import java.util.Objects;

public class Line {
    private final List<Point> points;

    private Line(List<Point> points) {
        this.points = points;
    }

    public static Line of(PointsStrategy pointsStrategy) {
        return new Line(pointsStrategy.points());
    }

    public int move(int point) {
        return points.get(point).move();
    }

    public int numberOfPoints() {
        return points.size();
    }

    public boolean hasRightLine(int point) {
        return points.get(point).isRight();
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Line line = (Line) o;
        return Objects.equals(points, line.points);
    }

    @Override
    public int hashCode() {
        return Objects.hash(points);
    }
}
